fix https://github.com/AlexAkulov/clickhouse-backup/issues/220, allow total_bytes as uint64 fetching fix allocations for `tableMetadataForDownload` fix getTableSizeFromParts behavior only for required tables fix Error handling on some suggested cases

+++ pkg/new_storage/general.go
@@ -96,7 +96,7 @@ func (bd *BackupDestination) BackupList() ([]Backup, error) {
 			result = append(result, Backup{
 				metadata.BackupMetadata{
 					BackupName: backupName,
-					DataSize:   o.Size(),
+					DataSize:   uint64(o.Size()),
 				},
 				true,
 				fileExtension,
